+++ wolfcrypt/src/sp_arm64.c
@@ -3107,7 +3107,7 @@ static int sp_2048_mod_exp_16(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 64)
             c = 60;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 16);
         for (; i>=0 || c>=4; ) {
@@ -3251,7 +3251,7 @@ static int sp_2048_mod_exp_16(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 64)
             c = 59;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 16);
         for (; i>=0 || c>=5; ) {
@@ -4475,7 +4475,7 @@ static int sp_2048_mod_exp_32(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 64)
             c = 60;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 32);
         for (; i>=0 || c>=4; ) {
@@ -4619,7 +4619,7 @@ static int sp_2048_mod_exp_32(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 64)
             c = 59;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 32);
         for (; i>=0 || c>=5; ) {
@@ -5281,7 +5281,7 @@ static int sp_2048_mod_exp_2_32(sp_digit* r, sp_digit* e, int bits,
         c -= bits % 6;
         if (c == 64)
             c = 58;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         sp_2048_lshift_32(r, norm, y);
         for (; i>=0 || c>=6; ) {
@@ -10100,7 +10100,7 @@ static int sp_3072_mod_exp_24(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 64)
             c = 60;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 24);
         for (; i>=0 || c>=4; ) {
@@ -10244,7 +10244,7 @@ static int sp_3072_mod_exp_24(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 64)
             c = 59;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 24);
         for (; i>=0 || c>=5; ) {
@@ -11836,7 +11836,7 @@ static int sp_3072_mod_exp_48(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 64)
             c = 60;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 48);
         for (; i>=0 || c>=4; ) {
@@ -11980,7 +11980,7 @@ static int sp_3072_mod_exp_48(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 64)
             c = 59;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 48);
         for (; i>=0 || c>=5; ) {
@@ -12738,7 +12738,7 @@ static int sp_3072_mod_exp_2_48(sp_digit* r, sp_digit* e, int bits,
         c -= bits % 6;
         if (c == 64)
             c = 58;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 64 - c;
         sp_3072_lshift_48(r, norm, y);
         for (; i>=0 || c>=6; ) {
