+++ wolfcrypt/src/sp_cortexm.c
@@ -3473,7 +3473,7 @@ static int sp_2048_mod_exp_32(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 32)
             c = 28;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 32);
         for (; i>=0 || c>=4; ) {
@@ -3617,7 +3617,7 @@ static int sp_2048_mod_exp_32(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 32);
         for (; i>=0 || c>=5; ) {
@@ -4154,7 +4154,7 @@ static int sp_2048_mod_exp_64(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 32)
             c = 28;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 64);
         for (; i>=0 || c>=4; ) {
@@ -4298,7 +4298,7 @@ static int sp_2048_mod_exp_64(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 64);
         for (; i>=0 || c>=5; ) {
@@ -5160,7 +5160,7 @@ static int sp_2048_mod_exp_2_64(sp_digit* r, sp_digit* e, int bits,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         sp_2048_lshift_64(r, norm, y);
         for (; i>=0 || c>=5; ) {
@@ -8510,7 +8510,7 @@ static int sp_3072_mod_exp_48(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 32)
             c = 28;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 48);
         for (; i>=0 || c>=4; ) {
@@ -8654,7 +8654,7 @@ static int sp_3072_mod_exp_48(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 48);
         for (; i>=0 || c>=5; ) {
@@ -9197,7 +9197,7 @@ static int sp_3072_mod_exp_96(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 32)
             c = 28;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 96);
         for (; i>=0 || c>=4; ) {
@@ -9341,7 +9341,7 @@ static int sp_3072_mod_exp_96(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 96);
         for (; i>=0 || c>=5; ) {
@@ -10399,7 +10399,7 @@ static int sp_3072_mod_exp_2_96(sp_digit* r, sp_digit* e, int bits,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         sp_3072_lshift_96(r, norm, y);
         for (; i>=0 || c>=5; ) {
